+++ db/seeds/projects.rb
@@ -1,0 +1,27 @@
+Project.create!(
+  title: "HeefusDeluxe",
+  description: "One of my favorite comedians, based right here in the Tacoma area, regularly performs across the Pacific Northwest and beyond. You can catch him hosting an open mic night at a local comedy club or otherwise performing a show near you! His sharp wit and hilarious takes keep the crowd entertained every time. If you're ever in the area, be sure to check out one of his shows and enjoy some of the best local comedy around!",
+  image: "project1.png", #referenced image filename within app/assets/images(do not include .png or other filetype)
+  url: "https://github.com/woximorks/heefusdeluxe"
+)
+
+Project.create!(
+  title: "GuacAMole",
+  description: "This is one of the first JavaScript projects I created while following a freeCodeCamp tutorial—though I decided to put my own spin on it! The original plan was to build a Whack-A-Mole game, but I ended up taking it in a different direction. It was a fun way to get creative while applying the core concepts I was learning, and it gave me the opportunity to experiment with JavaScript in a way that felt unique and personal.",
+  image: "project2.png", #referenced image filename within app/assets/images(do not include .png or other filetype)
+  url: "https://master.d2rge8c7efug7r.amplifyapp.com/"
+)
+
+Project.create!(
+  title: "Beard Bros Dumpsters",
+  description: "This Rails application was developed to support a Pennsylvania-based company that offers dumpster rental services, including delivery and junk removal assistance in the surrounding area. The app streamlines the booking process for customers while providing the business with essential tools to manage their operations efficiently. With a user-friendly interface, it allows residents and businesses alike to easily schedule rentals and junk removal, ensuring a seamless experience for both parties. The project reflects a commitment to enhancing service accessibility and operational effectiveness in the local community.",
+  image: "project3.png", #referenced image filename within app/assets/images(do not include .png or other filetype)
+  url: "https://www.beardbrosdumpsters.com/"
+)
+
+Project.create!(
+  title: "Close Enough Fabrication",
+  description: "This website was created to showcase one of my favorite hobbies outside of web development! It serves as a platform to share my creative endeavors through fabrication and related shop projects. You’ll find a collection of my latest work, detailing the processes and techniques I've used along the way. This site not only highlights my passion, but also allows me to connect with others who share similar interests through my instagram - @closeenoughfabrication!",
+  image: "project4.png", #referenced image filename within app/assets/images(do not include .png or other filetype)
+  url: "https://cefab-offroad.com/"
+)